Split test_linter_acceptance out of test_warnings
test_warnings finally grew to be too long, so its time to split
some of the unrelated tests out.
